requirements.txt
# TalentAI Solutions – Requirements
# Install: pip install -r requirements.txt

# Core data
pandas>=2.0.0
numpy>=1.24.0

# ML & NLP
scikit-learn>=1.3.0
scipy>=1.11.0
nltk>=3.8.0
python-dateutil>=2.8.2
joblib>=1.3.0

# API
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0

# Web app
streamlit>=1.32.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.13.0

# PDF extraction
PyMuPDF>=1.23.0     # fitz – used by text_extractor.py
# OR: pypdf>=3.0.0

# GenAI (optional – for AI Assistant & RAG Chatbot)
google-generativeai>=0.5.0
